--- client/src/styles/leaderboard.less
@@ -1,20 +1,0 @@
-#leaderboard {
-    position: absolute;
-    right: 10px;
-    top: 10px;
-    padding: 5px;
-    min-width: 100px;
-
-    background-color: rgb(64,64,64);
-    color: white;
-    font-size: 12px;
-
-    .entry {
-        display: flex;
-        flex-direction: row;
-        padding: 3px;
-        font-weight: bold;
-        font-size: 12px;
-        font-family: monospace;
-    }
-}
