+++ client/src/styles/main.less
@@ -76,4 +76,4 @@ button {
 @import "user-input.less";
 @import "snake-info.less";
 @import "fps-stats.less";
-@import "leaderboard.less";
+@import "game-statistics.less";
